+++ team_token_integration_test.go
@@ -5,6 +5,7 @@ package tfe
 
 import (
 	"context"
+	"fmt"
 	"testing"
 	"time"
 
@@ -92,6 +93,81 @@ func TestTeamTokens_CreateWithOptions(t *testing.T) {
 	})
 }
 
+func TestTeamTokens_CreateWithOptions_MultipleTokens(t *testing.T) {
+	skipUnlessBeta(t)
+
+	client := testClient(t)
+	ctx := context.Background()
+
+	tmTest, tmTestCleanup := createTeam(t, client, nil)
+	t.Cleanup(tmTestCleanup)
+
+	t.Run("with multiple tokens", func(t *testing.T) {
+		desc1 := fmt.Sprintf("go-tfe-team-token-test-%s", randomString(t))
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{
+			Description: desc1,
+		})
+		require.NoError(t, err)
+		require.NotEmpty(t, tt.Token)
+		require.Equal(t, tt.Description, desc1)
+
+		desc2 := fmt.Sprintf("go-tfe-team-token-test-%s", randomString(t))
+		tt, err = client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{
+			Description: desc2,
+		})
+		require.NoError(t, err)
+		require.NotEmpty(t, tt.Token)
+		require.Equal(t, tt.Description, desc2)
+	})
+
+	t.Run("with an expiration date", func(t *testing.T) {
+		desc := fmt.Sprintf("go-tfe-team-token-test-%s", randomString(t))
+		currentTime := time.Now().UTC().Truncate(time.Second)
+		oneDayLater := currentTime.Add(24 * time.Hour)
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{
+			Description: desc,
+			ExpiredAt:   &oneDayLater,
+		})
+		require.NoError(t, err)
+		require.NotEmpty(t, tt.Token)
+		assert.Equal(t, tt.ExpiredAt, oneDayLater)
+		require.Equal(t, tt.Description, desc)
+	})
+
+	t.Run("without an expiration date", func(t *testing.T) {
+		desc := fmt.Sprintf("go-tfe-team-token-test-%s", randomString(t))
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{
+			Description: desc,
+		})
+		require.NoError(t, err)
+		require.NotEmpty(t, tt.Token)
+		assert.Empty(t, tt.ExpiredAt)
+		require.Equal(t, tt.Description, desc)
+	})
+
+	t.Run("when a token already exists with the same description", func(t *testing.T) {
+		desc := fmt.Sprintf("go-tfe-team-token-test-%s", randomString(t))
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{
+			Description: desc,
+		})
+		require.NoError(t, err)
+		require.NotEmpty(t, tt.Token)
+		require.Equal(t, tt.Description, desc)
+
+		tt, err = client.TeamTokens.CreateWithOptions(ctx, tmTest.ID, TeamTokenCreateOptions{
+			Description: desc,
+		})
+		assert.Nil(t, tt)
+		assert.Equal(t, err, ErrInvalidDescriptionConflict)
+	})
+
+	t.Run("without valid team ID", func(t *testing.T) {
+		tt, err := client.TeamTokens.CreateWithOptions(ctx, badIdentifier, TeamTokenCreateOptions{})
+		assert.Nil(t, tt)
+		assert.Equal(t, err, ErrInvalidTeamID)
+	})
+}
+
 func TestTeamTokensRead(t *testing.T) {
 	client := testClient(t)
 	ctx := context.Background()
@@ -105,6 +181,8 @@ func TestTeamTokensRead(t *testing.T) {
 		tt, err := client.TeamTokens.Read(ctx, tmTest.ID)
 		require.NoError(t, err)
 		assert.NotEmpty(t, tt)
+		require.NotEmpty(t, tt.Team)
+		assert.Equal(t, tt.Team.ID, tmTest.ID)
 
 		ttTestCleanup()
 	})
@@ -117,8 +195,10 @@ func TestTeamTokensRead(t *testing.T) {
 
 		tt, err := client.TeamTokens.Read(ctx, tmTest.ID)
 		require.NoError(t, err)
-		assert.NotEmpty(t, tt)
+		require.NotEmpty(t, tt)
 		assert.Equal(t, tt.ExpiredAt, oneDayLater)
+		require.NotEmpty(t, tt.Team)
+		assert.Equal(t, tt.Team.ID, tmTest.ID)
 
 		ttTestCleanup()
 	})
@@ -133,6 +213,69 @@ func TestTeamTokensRead(t *testing.T) {
 		tt, err := client.OrganizationTokens.Read(ctx, badIdentifier)
 		assert.Nil(t, tt)
 		assert.EqualError(t, err, ErrInvalidOrg.Error())
+	})
+}
+
+func TestTeamTokensReadByID(t *testing.T) {
+	client := testClient(t)
+	ctx := context.Background()
+
+	tmTest, tmTestCleanup := createTeam(t, client, nil)
+	t.Cleanup(tmTestCleanup)
+
+	currentTime := time.Now().UTC().Truncate(time.Second)
+	oneDayLater := currentTime.Add(24 * time.Hour)
+	t.Run("with legacy, descriptionless tokens", func(t *testing.T) {
+		token, ttTestCleanup := createTeamTokenWithOptions(t, client, tmTest, TeamTokenCreateOptions{
+			ExpiredAt: &oneDayLater,
+		})
+		t.Cleanup(ttTestCleanup)
+
+		tt, err := client.TeamTokens.ReadByID(ctx, token.ID)
+		require.NoError(t, err)
+		require.NotEmpty(t, tt)
+		assert.Empty(t, tt.Description)
+		assert.Equal(t, tt.ExpiredAt, oneDayLater)
+		require.NotEmpty(t, tt.Team)
+		assert.Equal(t, tt.Team.ID, tmTest.ID)
+	})
+
+	t.Run("with multiple team tokens", func(t *testing.T) {
+		skipUnlessBeta(t)
+		desc1 := fmt.Sprintf("go-tfe-team-token-test-%s", randomString(t))
+		token, ttTestCleanup := createTeamTokenWithOptions(t, client, tmTest, TeamTokenCreateOptions{
+			Description: desc1,
+		})
+		t.Cleanup(ttTestCleanup)
+
+		tt, err := client.TeamTokens.ReadByID(ctx, token.ID)
+		require.NoError(t, err)
+		require.NotEmpty(t, tt)
+		assert.Equal(t, tt.Description, desc1)
+		assert.Empty(t, tt.ExpiredAt)
+		require.NotEmpty(t, tt.Team)
+		assert.Equal(t, tt.Team.ID, tmTest.ID)
+
+		desc2 := fmt.Sprintf("go-tfe-team-token-test-%s", randomString(t))
+		tokenWithExpiration, ttTestCleanup2 := createTeamTokenWithOptions(t, client, tmTest, TeamTokenCreateOptions{
+			ExpiredAt:   &oneDayLater,
+			Description: desc2,
+		})
+		t.Cleanup(ttTestCleanup2)
+
+		tt2, err := client.TeamTokens.ReadByID(ctx, tokenWithExpiration.ID)
+		require.NoError(t, err)
+		require.NotEmpty(t, tt2)
+		assert.Equal(t, tt2.Description, desc2)
+		assert.Equal(t, tt2.ExpiredAt, oneDayLater)
+		require.NotEmpty(t, tt.Team)
+		assert.Equal(t, tt.Team.ID, tmTest.ID)
+	})
+
+	t.Run("when a token doesn't exists", func(t *testing.T) {
+		tt, err := client.TeamTokens.ReadByID(ctx, "nonexistent-token-id")
+		assert.Equal(t, ErrResourceNotFound, err)
+		assert.Nil(t, tt)
 	})
 }
 
@@ -158,5 +301,46 @@ func TestTeamTokensDelete(t *testing.T) {
 	t.Run("without valid team ID", func(t *testing.T) {
 		err := client.TeamTokens.Delete(ctx, badIdentifier)
 		assert.Equal(t, err, ErrInvalidTeamID)
+	})
+}
+
+func TestTeamTokensDeleteByID(t *testing.T) {
+	client := testClient(t)
+	ctx := context.Background()
+
+	tmTest, tmTestCleanup := createTeam(t, client, nil)
+	t.Cleanup(tmTestCleanup)
+
+	t.Run("with legacy, descriptionless tokens", func(t *testing.T) {
+		token, _ := createTeamToken(t, client, tmTest)
+		err := client.TeamTokens.DeleteByID(ctx, token.ID)
+		require.NoError(t, err)
+	})
+
+	t.Run("with multiple team tokens", func(t *testing.T) {
+		skipUnlessBeta(t)
+		token1, _ := createTeamTokenWithOptions(t, client, tmTest, TeamTokenCreateOptions{
+			Description: fmt.Sprintf("go-tfe-team-token-test-%s", randomString(t)),
+		})
+
+		token2, _ := createTeamTokenWithOptions(t, client, tmTest, TeamTokenCreateOptions{
+			Description: fmt.Sprintf("go-tfe-team-token-test-%s", randomString(t)),
+		})
+
+		err := client.TeamTokens.DeleteByID(ctx, token1.ID)
+		require.NoError(t, err)
+
+		err = client.TeamTokens.DeleteByID(ctx, token2.ID)
+		require.NoError(t, err)
+	})
+
+	t.Run("when a token does not exist", func(t *testing.T) {
+		err := client.TeamTokens.DeleteByID(ctx, "nonexistent-token-id")
+		assert.Equal(t, err, ErrResourceNotFound)
+	})
+
+	t.Run("with invalid token ID", func(t *testing.T) {
+		err := client.TeamTokens.DeleteByID(ctx, badIdentifier)
+		assert.Equal(t, err, ErrInvalidTokenID)
 	})
 }